# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Manage Applications API",
        "description": "REST API documentation for the Manage Applications & Evaluators backend",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/v1/evaluators/{evaluatorId}/assign-candidate": {
            "post": {
                "tags": [
                    "Evaluators"
                ],
                "summary": "Assign a candidate to an evaluator",
                "operationId": "10d3dcaed0cad242ff37d5cb65943e3a",
                "parameters": [
                    {
                        "name": "evaluatorId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "candidate_id"
                                ],
                                "properties": {
                                    "candidate_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Candidate assigned successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Candidate assigned successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Evaluator or Candidate not found"
                    },
                    "409": {
                        "description": "Candidate already assigned"
                    }
                }
            }
        },
        "/api/v1/evaluators/{evaluatorId}/assignments/{candidateId}/complete": {
            "put": {
                "tags": [
                    "Evaluators"
                ],
                "summary": "Mark assignment as completed",
                "operationId": "ef0a819a4507fbad01f5be21902b1784",
                "parameters": [
                    {
                        "name": "evaluatorId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "candidateId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Assignment completed"
                    },
                    "404": {
                        "description": "Assignment not found"
                    }
                }
            }
        },
...
        "/api/v1/evaluators/{evaluatorId}/assignments/{candidateId}/reject": {
            "put": {
                "tags": [
                    "Evaluators"
                ],
                "summary": "Reject an assignment",
                "operationId": "f7afb62c42c0f1f6319497b490d57c6a",
                "parameters": [
                    {
                        "name": "evaluatorId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "candidateId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Assignment rejected"
                    },
                    "404": {
                        "description": "Assignment not found"
                    }
                }
            }
        },
...
        "/api/v1/evaluators/{evaluatorId}/assignments/{candidateId}/start-progress": {
            "put": {
                "tags": [
                    "Evaluators"
                ],
                "summary": "Move assignment to in_progress",
                "operationId": "1a26eb4faf45b4c579dd039241941683",
                "parameters": [
                    {
                        "name": "evaluatorId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "candidateId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Assignment moved to in_progress"
                    },
                    "404": {
                        "description": "Assignment not found"
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}